// NNetwork.Clients/nnetwork-react/src/enums/layer-type.ts
export enum LayerType
{
    Conv2D = "Conv2D",
    Dense = "Dense",
    Dropout = "Dropout",
    Flatten = "Flatten",
    Input = "Input",
    MaxPooling2D = "MaxPooling2D",
    UpSampling2D = "UpSampling2D",
    // Activation functions
    LeakyReLU = "LeakyReLU",
    PReLU = "PReLU",
    ReLU = "ReLU",
    ELU = "ELU",
    Softmax = "Softmax"
}